refactor(runtime): split orchestrator/tests — extract multimodal/attachments/resize/audio
Combine the next four related sections of the orchestrator test file
into a single `tests/multimodal.rs` (15 tests, ~430 lines):

MultimodalUser/OTel serialisation (3):
- `serialize_history_multimodal_user_omits_base64_data`
- `prepare_history_with_attachments_emits_multimodal_user`
- `prepare_history_without_attachments_emits_plain_text`

Attachment collection (5, including `MockAttachmentTool`):
- `run_turn_collects_attachments_from_tool_output`
- `run_turn_collects_multiple_attachments_across_tool_calls`
- `run_turn_no_attachments_when_tools_return_none`
- `run_turn_streaming_collects_attachments`
- `run_turn_with_tools_collects_attachments_from_extension`

resize_and_encode (5):
- `resize_and_encode_small_image_returns_base64_without_resize`
- `resize_and_encode_unknown_mime_returns_base64`
- `max_image_bytes_anthropic_is_5mb`
- `max_image_bytes_openai_is_20mb`
- `max_image_bytes_default_is_conservative`

AudioStore integration (2):
- `with_audio_store_makes_store_available`
- `audio_store_none_by_default`

The `build_with_executor` helper moves up into `tests/mod.rs`
alongside `build` since the `Slice C: turn_had_errors` tests still
living in mod.rs also use it.

cargo test -p assistant-runtime --lib reports 220 passed (same as
before); `orchestrator::tests::multimodal::*` subtotal is 15.

Fifth slice in the stack.
